+++ dash_extensions/_imports_.py
@@ -1,15 +1,15 @@
 from .Burger import Burger
-from .DashWebSocket import DashWebSocket
 from .Download import Download
 from .Keyboard import Keyboard
 from .Lottie import Lottie
 from .Monitor import Monitor
+from .WebSocket import WebSocket
 
 __all__ = [
     "Burger",
-    "DashWebSocket",
     "Download",
     "Keyboard",
     "Lottie",
-    "Monitor"
+    "Monitor",
+    "WebSocket"
 ]