+++ examples/websocket_clientside.py
@@ -1,7 +1,6 @@
 import dash_core_components as dcc
 import dash_html_components as html
 import dash_extensions as de
-
 from dash import Dash
 from dash.dependencies import Input, Output
 
@@ -9,7 +8,7 @@ from dash.dependencies import Input, Output
 app = Dash(prevent_initial_callbacks=True)
 app.layout = html.Div([
     dcc.Input(id="input", autoComplete="off"), html.Div(id="msg"),
-    de.DashWebSocket(url="wss://echo.websocket.org", id="ws")
+    de.WebSocket(url="wss://echo.websocket.org", id="ws")
 ])
 # Send input value using websocket.
 send = "function(value){return value;}"