+++ src/lib/index.js
@@ -4,7 +4,7 @@ import Lottie from './components/Lottie.react';
 import Keyboard from './components/Keyboard.react';
 import Monitor from './components/Monitor.react';
 import Burger from './components/Burger.react';
-import DashWebSocket from './components/DashWebSocket.react';
+import WebSocket from './components/WebSocket.react';
 
 export {
     Download,
@@ -12,5 +12,5 @@ export {
     Keyboard,
     Monitor,
     Burger,
-    DashWebSocket
+    WebSocket
 };